New source structure for multi platform
