+++ src/core/libutils.c
