+++ src/core/utils.c
