+++ src/devices/perf_device.c
@@ -1,0 +1,143 @@
+#include <stdlib.h>
+#include <sys/time.h>
+#include <assert.h>
+
+#include "libdevs.h"
+#include "devices/perf_device.h"
+#include "libutils.h"
+
+struct perf_device {
+	/* This must be the first field. See dev_pdev() for details. */
+	struct device		dev;
+
+	struct device		*shadow_dev;
+
+	uint64_t		read_count;
+	uint64_t		read_time_us;
+	uint64_t		write_count;
+	uint64_t		write_time_us;
+	uint64_t		reset_count;
+	uint64_t		reset_time_us;
+};
+
+static inline struct perf_device *dev_pdev(struct device *dev)
+{
+	return (struct perf_device *)dev;
+}
+
+static int pdev_read_blocks(struct device *dev, char *buf,
+		uint64_t first_pos, uint64_t last_pos)
+{
+	struct perf_device *pdev = dev_pdev(dev);
+	struct timeval t1, t2;
+	int rc;
+
+	assert(!gettimeofday(&t1, NULL));
+	rc = pdev->shadow_dev->read_blocks(pdev->shadow_dev, buf,
+		first_pos, last_pos);
+	assert(!gettimeofday(&t2, NULL));
+	pdev->read_count += last_pos - first_pos + 1;
+	pdev->read_time_us += diff_timeval_us(&t1, &t2);
+	return rc;
+}
+
+static int pdev_write_blocks(struct device *dev, const char *buf,
+		uint64_t first_pos, uint64_t last_pos)
+{
+	struct perf_device *pdev = dev_pdev(dev);
+	struct timeval t1, t2;
+	int rc;
+
+	assert(!gettimeofday(&t1, NULL));
+	rc = pdev->shadow_dev->write_blocks(pdev->shadow_dev, buf,
+		first_pos, last_pos);
+	assert(!gettimeofday(&t2, NULL));
+	pdev->write_count += last_pos - first_pos + 1;
+	pdev->write_time_us += diff_timeval_us(&t1, &t2);
+	return rc;
+}
+
+static int pdev_reset(struct device *dev)
+{
+	struct perf_device *pdev = dev_pdev(dev);
+	struct timeval t1, t2;
+	int rc;
+
+	assert(!gettimeofday(&t1, NULL));
+	rc = dev_reset(pdev->shadow_dev);
+	assert(!gettimeofday(&t2, NULL));
+	pdev->reset_count++;
+	pdev->reset_time_us += diff_timeval_us(&t1, &t2);
+	return rc;
+}
+
+static void pdev_free(struct device *dev)
+{
+	struct perf_device *pdev = dev_pdev(dev);
+	free_device(pdev->shadow_dev);
+}
+
+static const char *pdev_get_filename(struct device *dev)
+{
+	return dev_get_filename(dev_pdev(dev)->shadow_dev);
+}
+
+struct device *pdev_detach_and_free(struct device *dev)
+{
+	struct perf_device *pdev = dev_pdev(dev);
+	struct device *shadow_dev = pdev->shadow_dev;
+	pdev->shadow_dev = NULL;
+	pdev->dev.free = NULL;
+	free_device(&pdev->dev);
+	return shadow_dev;
+}
+
+struct device *create_perf_device(struct device *dev)
+{
+	struct perf_device *pdev;
+
+	pdev = malloc(sizeof(*pdev));
+	if (!pdev)
+		return NULL;
+
+	pdev->shadow_dev = dev;
+	pdev->read_count = 0;
+	pdev->read_time_us = 0;
+	pdev->write_count = 0;
+	pdev->write_time_us = 0;
+	pdev->reset_count = 0;
+	pdev->reset_time_us = 0;
+
+	pdev->dev.size_byte = dev->size_byte;
+	pdev->dev.block_order = dev->block_order;
+	pdev->dev.read_blocks = pdev_read_blocks;
+	pdev->dev.write_blocks = pdev_write_blocks;
+	pdev->dev.reset	= pdev_reset;
+	pdev->dev.free = pdev_free;
+	pdev->dev.get_filename = pdev_get_filename;
+
+	return &pdev->dev;
+}
+
+void perf_device_sample(struct device *dev,
+	uint64_t *pread_count, uint64_t *pread_time_us,
+	uint64_t *pwrite_count, uint64_t *pwrite_time_us,
+	uint64_t *preset_count, uint64_t *preset_time_us)
+{
+	struct perf_device *pdev = dev_pdev(dev);
+
+	if (pread_count)
+		*pread_count = pdev->read_count;
+	if (pread_time_us)
+		*pread_time_us = pdev->read_time_us;
+
+	if (pwrite_count)
+		*pwrite_count = pdev->write_count;
+	if (pwrite_time_us)
+		*pwrite_time_us = pdev->write_time_us;
+
+	if (preset_count)
+		*preset_count = pdev->reset_count;
+	if (preset_time_us)
+		*preset_time_us = pdev->reset_time_us;
+}
